Seasoning

+++ Assets/Scripts/preferences/StatPreference.cs
@@ -2,10 +2,10 @@ using static IngredientData;
 
 public class StatPreference : IPreferences{
     public IngredientStat stat;
-    public int min, max;
+    public float min, max;
 
     public bool valid(Meal m){
-        int current = m.getStats()[(int)stat];
+        float current = m.getStats()[(int)stat];
         return current >= min && current <= max;
     }
 }
